refactor: move parsing logic into a package
